+++ pyracf/__init__.py
@@ -2,8 +2,8 @@
 from .access.access_admin import AccessAdmin
 from .common.add_operation_error import AddOperationError
 from .common.alter_operation_error import AlterOperationError
-from .common.null_response_error import NullResponseError
 from .common.security_request_error import SecurityRequestError
+from .common.security_response_error import SecurityResponseError
 from .common.segment_error import SegmentError
 from .common.segment_trait_error import SegmentTraitError
 from .common.userid_error import UserIdError
