+++ pyracf/common/security_result.py
@@ -16,7 +16,6 @@ class SecurityResult:
         self.__result = XMLParser.fromstring(result_xml)
         self.__result_dictionary = {"securityResult": {}}
         self.__extract_results()
-        self.__result_dictionary = self.__check_result_type(self.__result_dictionary)
         if running_userid is not None:
             self.__result_dictionary["securityResult"][
                 "runningUserid"
@@ -64,15 +63,6 @@ class SecurityResult:
             info.append(item.text)
             self.__definition.remove(item)
 
-    def __check_result_type(self, result: dict) -> dict:
-        not_profiles = ["returnCode", "reasonCode", "runningUserid"]
-        for profile_type in result["securityResult"]:
-            if profile_type in not_profiles:
-                continue
-            if "error" in result["securityResult"][profile_type]:
-                return self.__organize_smo_error(result, profile_type)
-        return result
-
     def __extract_commands(self, filter_out_extra_messages: bool) -> None:
         """Extract commands section from XML into a list."""
         self.__definition_dictionary["commands"] = []
@@ -116,32 +106,6 @@ class SecurityResult:
                 self.__definition_dictionary["error"][item_tag] = int(item.text)
             except ValueError:
                 self.__definition_dictionary["error"][item_tag] = item.text
-
-    def __organize_smo_error(self, result: dict, profile_type: str) -> dict:
-        command_result = {}
-        command_result["safReturnCode"] = result["securityResult"][profile_type][
-            "error"
-        ]["errorFunction"]
-        command_result["returnCode"] = result["securityResult"][profile_type]["error"][
-            "errorCode"
-        ]
-        command_result["reasonCode"] = result["securityResult"][profile_type]["error"][
-            "errorReason"
-        ]
-        command_result["errorOffset"] = result["securityResult"][profile_type]["error"][
-            "errorOffset"
-        ]
-        command_result["textInError"] = result["securityResult"][profile_type]["error"][
-            "textInError"
-        ]
-        command_result["messages"] = [
-            result["securityResult"][profile_type]["error"]["errorMessage"],
-            "Please note that for any text in error,"
-            + " redacted values may skew offset calculations.",
-        ]
-        result["securityResult"][profile_type]["commands"] = [command_result]
-        del result["securityResult"][profile_type]["error"]
-        return result
 
     def __to_pascal_case(self, key: str) -> str:
         """Convert result dictionary keys to pascal case."""
