+++ tests/connection/test_connection_debug_logging.py
@@ -9,7 +9,7 @@ from unittest.mock import Mock, patch
 import __init__
 
 import tests.connection.test_connection_constants as TestConnectionConstants
-from pyracf import ConnectionAdmin, SecurityRequestError
+from pyracf import ConnectionAdmin, SecurityResponseError
 from pyracf.common.irrsmo00 import IRRSMO00
 
 # Resolves F401
@@ -60,7 +60,7 @@ class TestConnectionDebugLogging(unittest.TestCase):
                     "TESTGRP0",
                     traits=TestConnectionConstants.TEST_CONNECT_CONNECTION_REQUEST_TRAITS,
                 ),
-            except SecurityRequestError:
+            except SecurityResponseError:
                 pass
         error_log = self.ansi_escape.sub("", stdout.getvalue())
         self.assertEqual(
