+++ tests/test_runner.py
@@ -11,8 +11,8 @@ from tests.access.test_access_result_parser import TestAccessResultParser
 from tests.common.test_customize_segment_traits import TestCustomizeSegmentTraits
 from tests.common.test_install_precheck_script import TestInstallPrecheckScript
 from tests.common.test_logger import TestLogger
-from tests.common.test_null_response_error import TestNullResponseError
 from tests.common.test_run_as_userid import TestRunAsUserId
+from tests.common.test_security_request_error import TestSecurityRequestError
 from tests.connection.test_connection_debug_logging import TestConnectionDebugLogging
 from tests.connection.test_connection_request_builder import (
     TestConnectionRequestBuilder,
@@ -66,7 +66,7 @@ def __test_suite() -> unittest.TestSuite:
         TestCustomizeSegmentTraits,
         TestInstallPrecheckScript,
         TestLogger,
-        TestNullResponseError,
+        TestSecurityRequestError,
         TestRunAsUserId,
         TestConnectionResultParser,
         TestConnectionRequestBuilder,
